Allow more fine-grained control on how building placement should be done
Added `random_alternative` flag to indicate whether a random placement is selected if
the original target cannot be done (original behavior) or whether we should try to
find the nearest possible placement to the target (new behavior). Default is
to use random placement.

Also, the `placement_step` has been made configurable to support the above feature.