Tweak `next_expansion` identification to ignore locations with nearby town halls
The original behavior was to check if the pathing to the expansion is within the threshold,
which would indicate that it is taken. However, this only checks the CURRENT townhall being
checked. If the distant expansion location already has a town hall, the `find_placement` method
would find an alternative location nearby. This results to town halls being placed to one another in the expansion.

The condition has been changed to instead look for nearby town halls to indicate if it's already taken