Fix property_cache_once_per_frame
Current implementation of the "property_cache_once_per_frame" decorator
doesn't work with multiple instances of the same class. This is due to
the cache being stored to the function itself, rather than to the
individual object it's being called on.

Address this issue by adding a parameter to the decorator, which can
be used to define where the cache should be stored on the object, and
store the cache there.

This allows multiple bots to co-exist in a single process without
conflicting caches.